Module advent_solutions::advent2017::day04

Day 4: High-Entropy Passphrases

A new system policy has been put in place that requires all accounts to use a passphrase instead of simply a pass*word*. A passphrase consists of a series of words (lowercase letters) separated by spaces.

Functions

is_valid_part1
To ensure security, a valid passphrase must contain no duplicate words.

is_valid_part2
For added security, yet another system policy has been put in place. Now, a valid passphrase must contain no two words that are anagrams of each other - that is, a passphrase is invalid if any word's letters can be rearranged to form any other word in the passphrase.

parse_input

part1
The system's full passphrase list is available as your puzzle input. How many passphrases are valid?

part2
Under this new system policy, how many passphrases are valid?