Function advent_solutions::advent2017::day04::is_valid_part1

pub fn is_valid_part1(passphrase: &str) -> bool

To ensure security, a valid passphrase must contain no duplicate words.

For example:

aa bb cc dd ee is valid.

assert_eq!(is_valid_part1("aa bb cc dd ee"), true);

aa bb cc dd aa is not valid - the word aa appears more than once.

assert_eq!(is_valid_part1("aa bb cc dd aa"), false);

aa bb cc dd aaa is valid - aa and aaa count as different words.

assert_eq!(is_valid_part1("aa bb cc dd aaa"), true);