Function advent_solutions::advent2017::day09::part2

pub fn part2(root: &Node) -> usize

Now, you're ready to remove the garbage.

To prove you've removed it, you need to count all of the characters within the garbage. The leading and trailing < and > don't count, nor do any canceled characters or the ! doing the canceling.

<>, 0 characters.

let root = Node::from_bytes(b"<>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 0);

<random characters>, 17 characters.

let root = Node::from_bytes(b"<random characters>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 17);

<<<<>, 3 characters.

let root = Node::from_bytes(b"<<<<>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 3);

<{!>}>, 2 characters.

let root = Node::from_bytes(b"<{!>}>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 2);

<!!>, 0 characters.

let root = Node::from_bytes(b"<!!>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 0);

<!!!>>, 0 characters.

let root = Node::from_bytes(b"<!!!>>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 0);

<{o"i!a,<{i<a>, 10 characters.

let root = Node::from_bytes(b"<{o\"i!a,<{i<a>")
    .to_result()
    .unwrap();

assert_eq!(part2(&root), 10);

How many non-canceled characters are within the garbage in your puzzle input?