Module advent_solutions::advent2017::day16

Day 16: Permutation Promenade

You come upon a very unusual sight; a group of programs here appear to be dancing.

Enums

Move
The programs' dance consists of a sequence of dance moves:

Functions

parse_input

part1
You watch the dance for a while and record their dance moves (your puzzle input). In what order are the programs standing after their dance?

part2
Now that you're starting to get a feel for the dance moves, you turn your attention to the dance as a whole.