Function advent_solutions::advent2017::day16::part2

pub fn part2(moves: &[Move]) -> String

Now that you're starting to get a feel for the dance moves, you turn your attention to the dance as a whole.

Keeping the positions they ended up in from their previous dance, the programs perform it again and again: including the first dance, a total of one billion (1000000000) times.

In the example above, their second dance would begin with the order baedc, and use the same dance moves:

s1, a spin of size 1: cbaed.

x3/4, swapping the last two programs: cbade.

pe/b, swapping programs e and b: ceadb.

In what order are the programs standing after their billion dances?