Function advent_solutions::advent2017::day19::solve

pub fn solve(input: &str) -> (String, usize)

The little packet looks up at you, hoping you can help it find the way. What letters will it see (in the order it would see them) if it follows the path? (The routing diagram is very wide; make sure you view it without line wrapping.)

Part Two

The packet is curious how many steps it needs to go.

For example, using the same routing diagram from the example above...

     |
     |  +--+
     A  |  C
 F---|----E|--+
     |  |  |  D
     +B-+  +--+

...the packet would go:

6 steps down (including the first line at the top of the diagram).

3 steps right.

4 steps up.

3 steps right.

4 steps down.

3 steps right.

2 steps up.

13 steps left (including the F it stops on).

This would result in a total of 38 steps.

let (_, steps) = solve(&input);
assert_eq!(steps, 38);

How many steps does the packet need to go?