Module advent_solutions::advent2017::day21

Day 21: Fractal Art

You find a program trying to generate some art. It uses a strange process that involves repeatedly enhancing the detail of an image through a set of rules.

Structs

Grid
The image consists of a two-dimensional square grid of pixels that are either on (#) or off (.).

Functions

parse_input

part1
How many pixels stay on after 5 iterations?

part2
How many pixels stay on after 18 iterations?

solve
The program always begins with this pattern: