Module advent_solutions::advent2017::day23

Day 23: Coprocessor Conflagration

You decide to head directly to the CPU and fix the printer from there. As you get close, you find an experimental coprocessor doing so much work that the local programs are afraid it will halt and catch fire. This would cause serious issues for the rest of the computer, so you head in and see what you can do.

Enums

Instruction
The code it's running seems to be a variant of the kind you saw recently on that tablet. The general functionality seems very similar, but some of the instructions are different:

Value

Functions

is_prime

parse_input

part1
The coprocessor is currently set to some kind of debug mode, which allows for testing, but prevents it from doing any meaningful work.

part2
Now, it's time to fix the problem.